requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# PDF Generation
reportlab>=4.0.0
PyPDF2>=3.0.0

# Log Parsing
python-json-logger>=2.0.0
pyparsing>=3.0.0

# Network & Security
ipaddress>=1.0.23
requests>=2.31.0
dnspython>=2.4.0

# Threat Intelligence Integration
stix2>=3.0.1
taxii2-client>=2.3.0

# MITRE ATT&CK
# mitreattack-python>=3.0.0  # Optional - causes dependency conflicts

# Database
# sqlite3 is built-in to Python
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI Enhancement
click>=8.1.0
colorama>=0.4.6
rich>=13.0.0
tabulate>=0.9.0

# Utilities
tqdm>=4.65.0
pytz>=2023.3

# SIEM Integration
elasticsearch>=8.0.0
splunk-sdk>=1.7.0

# Ticketing Integration
# jira>=3.5.0  # Optional
# pyservicenow>=0.5.0  # Optional - causes dependency conflicts

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
